refactor: extract module
Move env related code into own module
