+++ src/utils/manifest.ts
@@ -2,7 +2,7 @@ import { PkgManifest } from "../types/global";
 import fs from "fs";
 import { assertIsError } from "./error-type-guards";
 import log from "../logger";
-import { env } from "../core";
+import { env } from "./env";
 
 /**
  * Attempts to load the manifest from the path specified in env
